More mesh refactoring, need to still upload matrices to the shader before 3d works.
Mesh system will be generic, so only one instance of each mesh file will be loaded. For procedural mesh class this means one unique (size, uv coords) instance of the procedural mesh will be handled by the mesh factory, still WIP and not working but rendering with no matrices works and renders the mesh as expected.